+++ src/_doc/vendored/mod.rs
@@ -15,13 +15,13 @@ crate::_doc_vendor! { mod: "argv", argv }
 crate::_doc_vendor! { mod: "bytehound-preload", bytehound_preload }
 crate::_doc_vendor! { mod: "crossbeam-utils", crossbeam_utils }
 crate::_doc_vendor! { mod: "etcetera", etcetera}
-crate::_doc_vendor! { mod: "fxhash", fxhash }
 crate::_doc_vendor! { mod: "macro_rules_attribute", macro_rules_attribute }
 crate::_doc_vendor! { mod: "mini-alloc", mini_alloc }
 crate::_doc_vendor! { mod: "no_std_io", no_std_io }
 crate::_doc_vendor! { mod: "pastey", pastey }
 crate::_doc_vendor! { mod: "pollster", pollster }
 crate::_doc_vendor! { mod: "quickdiv", quickdiv }
+crate::_doc_vendor! { mod: "rustc-hash", rustc_hash }
 crate::_doc_vendor! { mod: "stack_dst", stack_dst }
 crate::_doc_vendor! { mod: "stated-scope-guard", stated_scope_guard }
 crate::_doc_vendor! { mod: "static_assertions", static_assertions }